Ateraçãoes no front end botões deletar e editar

+++ src/controllers/paletas.controller.js
@@ -7,6 +7,9 @@ const findPaletasController = (req, res) => {
 
 const findPaletaByIdController = (req, res) => {
   const idParam = req.params.id;
+  if (!idParam) {
+    return res.status(404).send({ message: "Paleta não encontrada!" })
+  }
   const chosenPaleta = paletasService.findPaletaByIdService(idParam);
   res.send(chosenPaleta);
 };
@@ -17,21 +20,31 @@ const createPaletaController = (req, res) => {
     !paleta.sabor ||
     !paleta.descricao ||
     !paleta.foto ||
-    !paleta.preco
+    !paleta.preco 
   ) {
-    return res.send('Você não preencheu todos os dados para adicionar uma nova paleta ao cardápio!',);}
+    res.status(400).send({ mensagem: "Você não preencheu todos os dados para adicionar uma nova paleta ao cardápio!" });
+  }
     const newPaleta = paletasService.createPaletaService(paleta);
     res.send(newPaleta);
 };
 const updatePaletaController = (req, res) => {
   const idParam = +req.params.id;
   const paletaEdit = req.body;
+  if (!idParam) {
+    return res.status(404).send({ message: "Paleta não encontrada!" })
+  }
+  if (!paletaEdit || !paletaEdit.sabor || !paletaEdit.descricao || !paletaEdit.foto || !paletaEdit.preco) {
+    return res.status(400).send({ message: "Você não preencheu todos os dados para editar a paleta!" });
+  }
   const updatedPaleta = paletasService.updatePaletaService(idParam, paletaEdit);
   res.send(updatedPaleta);
 };
 
 const deletePaletaController = (req, res) => {
   const idParam = req.params.id;
+  if (!idParam) {
+    return res.status(404).send({ message: "Paleta não encontrada!" })
+  }
   paletasService.deletePaletaService(idParam);
   res.send({ message: 'Paleta deletada com sucesso!' });
 };
